Switch content_size from int to string in SafeBrowsingPrivateEventRouter
Bug: 1293855
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3437113
Commit-Queue: Dominique Fauteux-Chapleau <domfc@chromium.org>
Cr-Commit-Position: refs/heads/main@{#967970}

+++ chrome/browser/safe_browsing/cloud_content_scanning/deep_scanning_test_utils.h
@@ -45,7 +45,7 @@ class EventReportValidator {
       const std::string& expected_threat_type,
       const std::string& expected_trigger,
       const std::set<std::string>* expected_mimetypes,
-      int expected_content_size,
+      int64_t expected_content_size,
       const std::string& expected_result,
       const std::string& expected_username,
       const absl::optional<std::string>& expected_scan_id);
@@ -58,7 +58,7 @@ class EventReportValidator {
       const enterprise_connectors::ContentAnalysisResponse::Result&
           expected_dlp_verdict,
       const std::set<std::string>* expected_mimetypes,
-      int expected_content_size,
+      int64_t expected_content_size,
       const std::string& expected_result,
       const std::string& expected_username,
       const std::string& expected_scan_id);
@@ -71,7 +71,7 @@ class EventReportValidator {
       const std::vector<enterprise_connectors::ContentAnalysisResponse::Result>&
           expected_dlp_verdicts,
       const std::set<std::string>* expected_mimetypes,
-      int expected_content_size,
+      int64_t expected_content_size,
       const std::vector<std::string>& expected_results,
       const std::string& expected_username,
       const std::vector<std::string>& expected_scan_ids);
@@ -85,7 +85,7 @@ class EventReportValidator {
       const enterprise_connectors::ContentAnalysisResponse::Result&
           expected_dlp_verdict,
       const std::set<std::string>* expected_mimetypes,
-      int expected_content_size,
+      int64_t expected_content_size,
       const std::string& expected_result,
       const std::string& expected_username,
       const std::string& expected_scan_id);
@@ -99,7 +99,7 @@ class EventReportValidator {
       const enterprise_connectors::ContentAnalysisResponse::Result&
           expected_dlp_verdict,
       const std::set<std::string>* expected_mimetypes,
-      int expected_content_size,
+      int64_t expected_content_size,
       const std::string& expected_result,
       const std::string& expected_username,
       const std::string& expected_scan_id);
@@ -110,7 +110,7 @@ class EventReportValidator {
                                 const std::string& expected_trigger,
                                 const std::string& expected_reason,
                                 const std::set<std::string>* expected_mimetypes,
-                                int expected_content_size,
+                                int64_t expected_content_size,
                                 const std::string& expected_result,
                                 const std::string& expected_username);
 
@@ -121,7 +121,7 @@ class EventReportValidator {
       const std::string& expected_trigger,
       const std::string& expected_reason,
       const std::set<std::string>* expected_mimetypes,
-      int expected_content_size,
+      int64_t expected_content_size,
       const std::string& expected_result,
       const std::string& expected_username);
 
@@ -132,7 +132,7 @@ class EventReportValidator {
       const std::string& expected_threat_type,
       const std::string& expected_trigger,
       const std::set<std::string>* expected_mimetypes,
-      int expected_content_size,
+      int64_t expected_content_size,
       const std::string& expected_result,
       const std::string& expected_username,
       const absl::optional<std::string>& expected_scan_id);
@@ -186,7 +186,7 @@ class EventReportValidator {
   absl::optional<std::string> trigger_ = absl::nullopt;
   absl::optional<std::string> threat_type_ = absl::nullopt;
   absl::optional<std::string> unscanned_reason_ = absl::nullopt;
-  absl::optional<int> content_size_ = absl::nullopt;
+  absl::optional<int64_t> content_size_ = absl::nullopt;
   raw_ptr<const std::set<std::string>> mimetypes_ = nullptr;
   std::string username_;
   absl::optional<bool> is_federated_ = absl::nullopt;
